place solitary physfacet and dimensions tags in proper notes
Modified the physdesc customizations to allow physfacet and dimensions tags that appear without a proper "extent" tag to be placed in "physfacet" or "dimensions" notes.

Currently unresolved:
I'd like to also be able to account for occasions where container summaries appear without a true extent, as well as have some catch-all option for any non-standard tags that may appear in a physdesc statement. I've attempted this by trying to throw all of those tags into new physdesc notes, but the importer crashes when the code runs and I haven't been able to figure out why. The attempted code is in comments towards the bottom of the physdesc section. Any ideas as to why that would be crashing?