fix for faulty parsing of primary language of material
By default, ASpace just uses the last <language> tag it finds as the primary language of the material described. This results in incorrect finding-aid languages for many eads.

For example, ead with the following <langmaterial> tag:

```xml
<langmaterial>
   The material is mostly in <language langcode="eng" encodinganalog="041">English</language>; 
   some correspondence is in <language langcode="arm" encodinganalog="041">Armenian;</language>; 
   select items are in <language langcode="ger" encodinganalog="041">German</language>.
</langmaterial>
```

will result in a primary material language of German. This should fix that.